add fit results for WW in ele/mu
So, that's how a Fighter plays, for me, at least in the early game. The mid-game is quite similar to what we've been doing for the last few zones -- Shield Pummel, Assault, use Step Up to reposition (either adjacent to the next enemy we want to kill, or far away if we want to recuperate).

We haven't maxed Shattering Shout yet; it grows to a huge radius 8 cone doing hundreds of damage. That will be our next priority for class talent points, since we have very little ranged damage capability right now.

Apart from that, the essential elements of our build are in place. We've maxed Shield Wall, Armour Training, and Step Up, which gives us incredible defense and mobility. We use selective offensive talents together with our mobility to kill targets of choice, and for less threatening opponents, we stand our ground and simply "bump" them to death.

I hope you've enjoyed this walkthrough and learned something from it. I did not avoid spoiling anything, so there are some fairly spoiler-ish hints scattered throughout this walkthrough, and even experienced players may pick up a few pointers.

This is not intended to be The One And Only Way To Build A Fighter by any means. It's just a build that I've used several times and found to be consistently strong. You should experiment with your own builds, and see what works and what doesn't. You may even turn out to be better at this than I am! (That wouldn't be terribly hard....)

Above all, enjoy!

HP/LP separately